Student Deatails shown

+++ rimdm_backend/routes/web.php
@@ -25,18 +25,20 @@ Route::get('/admin/{dashboard?}', [DashboardController::class, 'index'])
     ->name('dashboard')->middleware('admin');
 
 Route::prefix('/teacher')->group(function(){
+
     Route::get('/login', [AuthenticationController::class, 'login'])->name('teacher.login');
     Route::post('/loginPost', [AuthenticationController::class, 'loginPost'])->name('teacher.loginPost');
     Route::get('/logout', [AuthenticationController::class, 'logout'])->name('teacher.logout');
     Route::get('/profile', function () {return view('teacher.profile');})->name('teacher.profile')->middleware('teacher');
 
 
-    Route::get('/create/student', [StudentController::class, 'create'])->name('student.create');
-    Route::post('/store/student', [StudentController::class, 'store'])->name('student.store');
-    Route::get('/students', [StudentController::class, 'index'])->name('student.index');
-    Route::delete('/students/{id}', [StudentController::class, 'destroy'])->name('student.destroy');
-    Route::put('/students/{id}', [StudentController::class, 'update'])->name('student.update');
-    Route::get('/students/{id}/edit', [StudentController::class, 'edit'])->name('student.edit');
+    Route::get('/create/student', [StudentController::class, 'create'])->name('student.create')->middleware('admin');
+    Route::post('/store/student', [StudentController::class, 'store'])->name('student.store')->middleware('admin');
+    Route::get('/students', [StudentController::class, 'index'])->name('student.index')->middleware('admin');
+    Route::delete('/students/{id}', [StudentController::class, 'destroy'])->name('student.destroy')->middleware('admin');
+    Route::put('/students/{id}', [StudentController::class, 'update'])->name('student.update')->middleware('admin');
+    Route::get('/students/{id}/edit', [StudentController::class, 'edit'])->name('student.edit')->middleware('admin');
+    Route::get('/students/{id}', [StudentController::class, 'show'])->name('student.show')->middleware('admin');
 
 
 });
